backend/requirements.txt
# Aligned with .cursor/rules/project.mdc: FastAPI + Supabase only
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
openai>=1.0.0
anthropic>=0.18.0
supabase>=2.0.0
PyJWT[crypto]>=2.8.0
tiktoken>=0.5.0
psycopg[binary]>=3.1.0